backend/ingestors/mea_scraper.py
"""MEA Website Scraper - Data Ingestion"""

import logging
from datetime import datetime
from typing import List, Dict, Optional
import requests
from bs4 import BeautifulSoup
from urllib.parse import urljoin

from config import settings

logger = logging.getLogger(__name__)


class MEAScraper:
    """Web scraper for Ministry of External Affairs data"""
    
    def __init__(self):
        self.base_url = settings.MEA_BASE_URL
        self.headers = {
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
        }
    
    async def fetch_country_relations(self) -> List[Dict]:
        """
        Fetch bilateral relations data from MEA website
        
        Returns:
            List of country relation dictionaries
        """
        relations = []
        
        try:
            url = f"{self.base_url}/foreign-relations.htm"
            logger.info(f"Fetching MEA data from {url}")
            
            response = requests.get(url, headers=self.headers, timeout=10)
            response.raise_for_status()
            
            soup = BeautifulSoup(response.content, 'html.parser')
            
            # Extract country links and data
            # This would need to be customized based on actual MEA website structure
            country_elements = soup.find_all('a', {'class': 'country-link'})
            
            for element in country_elements:
                country_name = element.text.strip()
                country_url = urljoin(self.base_url, element.get('href'))
                
                relation_data = await self._scrape_country_detail(country_name, country_url)
                if relation_data:
                    relations.append(relation_data)
            
            logger.info(f"Successfully fetched {len(relations)} country relations")
            return relations
            
        except requests.RequestException as e:
            logger.error(f"Error fetching MEA data: {e}")
            return []
    
    async def _scrape_country_detail(self, country_name: str, url: str) -> Optional[Dict]:
        """
        Scrape detailed information for a specific country
        
        Args:
            country_name: Name of the country
            url: URL to scrape
            
        Returns:
            Dictionary with country relation data
        """
        try:
            response = requests.get(url, headers=self.headers, timeout=10)
            response.raise_for_status()
            
            soup = BeautifulSoup(response.content, 'html.parser')
            
            # Extract specific data points
            trade_volume = self._extract_trade_volume(soup)
            agreements = self._extract_agreements(soup)
            key_issues = self._extract_key_issues(soup)
            sentiment = self._analyze_sentiment(soup.get_text())
            
            return {
                "country": country_name,
                "relation_type": "bilateral",
                "status": "stable",  # This would be determined from content
                "trade_volume": trade_volume,
                "agreements": agreements,
                "key_issues": key_issues,
                "sentiment": sentiment,
                "confidence_score": 0.85,
                "last_updated": datetime.utcnow().isoformat(),
                "source": "MEA"
            }
            
        except requests.RequestException as e:
            logger.error(f"Error scraping {country_name}: {e}")
            return None
    
    def _extract_trade_volume(self, soup: BeautifulSoup) -> Optional[float]:
        """Extract trade volume from HTML"""
        # Implementation would depend on actual HTML structure
        return None
    
    def _extract_agreements(self, soup: BeautifulSoup) -> List[str]:
        """Extract list of agreements from HTML"""
        agreements = []
        # Implementation would depend on actual HTML structure
        return agreements
    
    def _extract_key_issues(self, soup: BeautifulSoup) -> List[str]:
        """Extract key issues from HTML"""
        issues = []
        # Implementation would depend on actual HTML structure
        return issues
    
    def _analyze_sentiment(self, text: str) -> str:
        """
        Simple sentiment analysis
        
        Args:
            text: Text to analyze
            
        Returns:
            Sentiment label: 'positive', 'neutral', 'negative'
        """
        # Placeholder - would use spaCy or TextBlob in production
        return "neutral"


async def run_mea_scraper():
    """Execute MEA scraping job"""
    scraper = MEAScraper()
    relations = await scraper.fetch_country_relations()
    
    # TODO: Save to Kafka and PostgreSQL
    logger.info(f"Extracted {len(relations)} country relations")
    return relations
