backend/requirements.txt
# ONTORA Backend - CORE PRODUCTION DEPENDENCIES
# Minimal set for Docker deployment
# See requirements-dev.txt for development tools
# See requirements-optional.txt for heavy packages

# Web Framework & Async
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & ORM
sqlalchemy==2.0.48
asyncpg==0.29.0
# Note: psycopg2-binary moved to requirements-windows.txt (Windows build issues)
neo4j==5.20.0
redis==5.0.1

# Data Processing (Lite)
numpy>=2.0,<2.3.0
pandas==2.1.3

# HTTP & Web
requests==2.31.0
httpx==0.25.2
beautifulsoup4==4.12.2

# Messaging
kafka-python==2.0.2

# Utilities & Security
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Monitoring
prometheus-client==0.19.0
python-json-logger==2.0.7

